Add SQS FIFO support
**Why was this change necessary?**
When using the register CLI to register an SQS FIFO destination queue,
you get the error "Invalid url, must be a fully qualified Amazon SQS
Queue Url".

**How does it address the problem?**
The regex in the checkSqsTarget method does not take into account that
FIFO queues must have .fifo at the end of them. This commit adds that as
a possibility to the regex.